Port the network environment validation to Ansible
This moves the code from the `network-environment-validator.py` under
`ansible-tests/validations/library/validate_network_environment.py` and
turn it into an Ansible module.

We add a validation that calls that module as well as update the
original validator script to load the same library code. That way, the
logic is in one place.